Split typed feedback guard helpers
