+++ crates/perry-runtime/src/typed_feedback/tests.rs
@@ -15,10 +15,7 @@ extern "C" fn test_direct_method(_this: f64, value: f64) -> f64 {
     value
 }
 
-extern "C" fn test_direct_closure(
-    _closure: *const crate::closure::ClosureHeader,
-    arg: f64,
-) -> f64 {
+extern "C" fn test_direct_closure(_closure: *const crate::closure::ClosureHeader, arg: f64) -> f64 {
     arg
 }
 
@@ -432,8 +429,7 @@ fn typed_feedback_class_field_set_guard_fails_for_frozen_object() {
     crate::object::js_object_set_field(obj, 0, crate::JSValue::from_bits(1.0f64.to_bits()));
     crate::object::js_object_freeze(receiver);
 
-    let guard =
-        js_typed_feedback_class_field_set_guard(31, receiver, class_id, keys, key, 0, 2.0);
+    let guard = js_typed_feedback_class_field_set_guard(31, receiver, class_id, keys, key, 0, 2.0);
     assert_eq!(guard, 0);
     assert_eq!(
         crate::object::js_object_get_field(obj, 0).bits(),
@@ -466,8 +462,7 @@ fn typed_feedback_class_field_set_guard_falls_back_for_class_setter() {
         );
     }
 
-    let guard =
-        js_typed_feedback_class_field_set_guard(32, receiver, class_id, keys, key, 0, 7.0);
+    let guard = js_typed_feedback_class_field_set_guard(32, receiver, class_id, keys, key, 0, 7.0);
     assert_eq!(guard, 0);
     js_typed_feedback_record_fallback_call(32);
     crate::object::js_object_set_field_by_name(obj, key, 7.0);
